fix nOctup arithmetic (#68)
* fix nOctup arithmetic

Fix provided by @JN-Jones as described in https://github.com/elucideye/acf/issues/67

* remove logging message

* #!/bin/bash -e

don’t ignore errors during setup in hunter_env.sh

+++ src/lib/acf/chnsPyramid.cpp
@@ -446,7 +446,7 @@ int Detector::getScales(int nPerOct, int nOctUp, const cv::Size& minDs, int shri
 
     for (int i = 0; i < nScales; i++)
     {
-        double s = std::pow(2.0, -double(i) / double(nPerOct + nOctUp));
+        double s = std::pow(2.0, -double(i) / double(nPerOct) + double(nOctUp));
         double s0 = (util::round(d0 * s / shrink) * shrink - 0.25 * shrink) / d0;
         double s1 = (util::round(d0 * s / shrink) * shrink + 0.25 * shrink) / d0;
         std::pair<double, double> best(0, std::numeric_limits<double>::max());
